feat: remove matomo add fathom
