+++ apps/website/project.json
@@ -33,6 +33,7 @@
       "defaultConfiguration": "development",
       "options": {
         "buildTarget": "website:build",
+        "hostname": "0.0.0.0",
         "port": 3001
       },
       "configurations": {
