+++ content/example/rss/feed.xml
@@ -4,7 +4,7 @@
         <title>Sia Foundation</title>
         <link>https://sia.tech</link>
         <description></description>
-        <lastBuildDate>Tue, 18 Jul 2023 20:04:48 GMT</lastBuildDate>
+        <lastBuildDate>Fri, 21 Jul 2023 13:45:56 GMT</lastBuildDate>
         <docs>https://validator.w3.org/feed/docs/rss2.html</docs>
         <generator>Feed for Node.js</generator>
         <image>
@@ -25,9 +25,9 @@
 <p>David Vorick, CEO of Skynet Labs (formerly Nebulous), says this will enable Skynet Labs to capitalize on the platform&#x27;s recent growth. “There has always been a tension in the Sia ecosystem between what the community needs and what the for-profit entity needs. Now that there is a nonprofit foundation that is solely accountable to the community, everyone will be better positioned to thrive.”</p>
 <p>The Sia Foundation is funded by a subsidy of Siacoins, added to the blockchain by a &quot;hardfork&quot; software update that activated last night. The update was adopted by a majority of the network&#x27;s users, many of whom expressed public support for the move. According to community member and contributor, SiaStats, the Sia Foundation “adds value and benefits every actor on the network. It safeguards the mission of Sia and enables grants, new developers, marketing, and growth actions.”</p>
 <p>In keeping with its commitment to improving the Sia network, in 2021 the Foundation plans to undertake a substantial overhaul of the Sia consensus algorithm, allowing the software to be used on smartphones and other low-power devices and enabling more users to fully validate the Sia blockchain. As for Skynet Labs, the company aims to replace modern web infrastructure with a decentralized upgrade which allows users to have control over their data and allows developers to innovate more rapidly.</p>
-<div><a id="about-the-sia-foundation" href="#about-the-sia-foundation"><span><svg focusable="false" preserveAspectRatio="xMidYMid meet" xmlns="http://www.w3.org/2000/svg" fill="currentColor" width="20" height="20" viewBox="0 0 32 32" aria-hidden="true"><path d="M29.25,6.76a6,6,0,0,0-8.5,0l1.42,1.42a4,4,0,1,1,5.67,5.67l-8,8a4,4,0,1,1-5.67-5.66l1.41-1.42-1.41-1.42-1.42,1.42a6,6,0,0,0,0,8.5A6,6,0,0,0,17,25a6,6,0,0,0,4.27-1.76l8-8A6,6,0,0,0,29.25,6.76Z"></path><path d="M4.19,24.82a4,4,0,0,1,0-5.67l8-8a4,4,0,0,1,5.67,0A3.94,3.94,0,0,1,19,14a4,4,0,0,1-1.17,2.85L15.71,19l1.42,1.42,2.12-2.12a6,6,0,0,0-8.51-8.51l-8,8a6,6,0,0,0,0,8.51A6,6,0,0,0,7,28a6.07,6.07,0,0,0,4.28-1.76L9.86,24.82A4,4,0,0,1,4.19,24.82Z"></path></svg></span><span as="h3">About the Sia Foundation</span></a></div>
+<div><div><div id="about-the-sia-foundation" class="absolute -top-[100px]"></div></div><a href="#about-the-sia-foundation"><span><svg focusable="false" preserveAspectRatio="xMidYMid meet" xmlns="http://www.w3.org/2000/svg" fill="currentColor" width="20" height="20" viewBox="0 0 32 32" aria-hidden="true"><path d="M29.25,6.76a6,6,0,0,0-8.5,0l1.42,1.42a4,4,0,1,1,5.67,5.67l-8,8a4,4,0,1,1-5.67-5.66l1.41-1.42-1.41-1.42-1.42,1.42a6,6,0,0,0,0,8.5A6,6,0,0,0,17,25a6,6,0,0,0,4.27-1.76l8-8A6,6,0,0,0,29.25,6.76Z"></path><path d="M4.19,24.82a4,4,0,0,1,0-5.67l8-8a4,4,0,0,1,5.67,0A3.94,3.94,0,0,1,19,14a4,4,0,0,1-1.17,2.85L15.71,19l1.42,1.42,2.12-2.12a6,6,0,0,0-8.51-8.51l-8,8a6,6,0,0,0,0,8.51A6,6,0,0,0,7,28a6.07,6.07,0,0,0,4.28-1.76L9.86,24.82A4,4,0,0,1,4.19,24.82Z"></path></svg></span><span as="h3">About the Sia Foundation</span></a></div>
 <p>The Sia Foundation is a non-profit corporation that maintains, improves, and promotes the <a rel="noopener" target="_blank" href="https://sia.tech">Sia</a> distributed storage network. It was founded in 2020 and is funded by a perpetual subsidy built into the Sia protocol. All code, research, educational materials, and other products of the Foundation&#x27;s work will be made available to the public free of charge and under open-source licenses.</p>
-<div><a id="about-skynet-labs" href="#about-skynet-labs"><span><svg focusable="false" preserveAspectRatio="xMidYMid meet" xmlns="http://www.w3.org/2000/svg" fill="currentColor" width="20" height="20" viewBox="0 0 32 32" aria-hidden="true"><path d="M29.25,6.76a6,6,0,0,0-8.5,0l1.42,1.42a4,4,0,1,1,5.67,5.67l-8,8a4,4,0,1,1-5.67-5.66l1.41-1.42-1.41-1.42-1.42,1.42a6,6,0,0,0,0,8.5A6,6,0,0,0,17,25a6,6,0,0,0,4.27-1.76l8-8A6,6,0,0,0,29.25,6.76Z"></path><path d="M4.19,24.82a4,4,0,0,1,0-5.67l8-8a4,4,0,0,1,5.67,0A3.94,3.94,0,0,1,19,14a4,4,0,0,1-1.17,2.85L15.71,19l1.42,1.42,2.12-2.12a6,6,0,0,0-8.51-8.51l-8,8a6,6,0,0,0,0,8.51A6,6,0,0,0,7,28a6.07,6.07,0,0,0,4.28-1.76L9.86,24.82A4,4,0,0,1,4.19,24.82Z"></path></svg></span><span as="h3">About Skynet Labs</span></a></div>
+<div><div><div id="about-skynet-labs" class="absolute -top-[100px]"></div></div><a href="#about-skynet-labs"><span><svg focusable="false" preserveAspectRatio="xMidYMid meet" xmlns="http://www.w3.org/2000/svg" fill="currentColor" width="20" height="20" viewBox="0 0 32 32" aria-hidden="true"><path d="M29.25,6.76a6,6,0,0,0-8.5,0l1.42,1.42a4,4,0,1,1,5.67,5.67l-8,8a4,4,0,1,1-5.67-5.66l1.41-1.42-1.41-1.42-1.42,1.42a6,6,0,0,0,0,8.5A6,6,0,0,0,17,25a6,6,0,0,0,4.27-1.76l8-8A6,6,0,0,0,29.25,6.76Z"></path><path d="M4.19,24.82a4,4,0,0,1,0-5.67l8-8a4,4,0,0,1,5.67,0A3.94,3.94,0,0,1,19,14a4,4,0,0,1-1.17,2.85L15.71,19l1.42,1.42,2.12-2.12a6,6,0,0,0-8.51-8.51l-8,8a6,6,0,0,0,0,8.51A6,6,0,0,0,7,28a6.07,6.07,0,0,0,4.28-1.76L9.86,24.82A4,4,0,0,1,4.19,24.82Z"></path></svg></span><span as="h3">About Skynet Labs</span></a></div>
 <p><a rel="noopener" target="_blank" href="https://siasky.net">Skynet Labs</a>, formerly known as Nebulous, builds uncompromising software infrastructure for the decentralized internet and is the team responsible for Skynet. Skynet is an application development and hosting platform and storage on Skynet is provided by Sia, the leading decentralized cloud storage network.</p>
 <p>Skynet Labs defines uncompromising infrastructure as scalable, trustless, secure, and – most importantly – fully decentralized. In a blockchain industry filled with hype but lacking substance, Skynet Labs stands out as one of the few deeply technical teams that consistently delivers real products with significant potential.</p>
 <p>Nebulous, Inc. was founded in 2014 and is funded by Paradigm, Bain Capital Ventures, A.Capital, Bessemer Venture Partners, First Star Ventures, and other notable investors.</p>]]></content:encoded>
