+++ libs/design-system/src/site/SitePageHead.tsx
@@ -51,17 +51,6 @@ export function SitePageHead({
         <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#1ed660" />
         <meta name="msapplication-TileColor" content="#2b5797" />
         <meta name="theme-color" content="#ffffff" />
-        {analytics && (
-          <noscript>
-            {/* eslint-disable-next-line */}
-            <img
-              referrerPolicy="no-referrer-when-downgrade"
-              src="https://surveillance.sia.tech/matomo.php?idsite=1&amp;rec=1"
-              style={{ border: 0 }}
-              alt=""
-            />
-          </noscript>
-        )}
         {children}
         <title>{fullTitle}</title>
         <meta name="description" content={description}></meta>
@@ -80,23 +69,10 @@ export function SitePageHead({
       </Head>
       {analytics && (
         <Script
-          id="matomo"
+          id="fathom"
           strategy="afterInteractive"
-          dangerouslySetInnerHTML={{
-            __html: `
-              var _paq = window._paq = window._paq || [];
-              /* tracker methods like "setCustomDimension" should be called before "trackPageView" */
-              _paq.push(['trackPageView']);
-              _paq.push(['enableLinkTracking']);
-              (function() {
-                var u="//surveillance.sia.tech/";
-                _paq.push(['setTrackerUrl', u+'matomo.php']);
-                _paq.push(['setSiteId', '1']);
-                var d=document, g=d.createElement('script'), s=d.getElementsByTagName('script')[0];
-                g.async=true; g.src=u+'matomo.js'; s.parentNode.insertBefore(g,s);
-              })();
-              `,
-          }}
+          src="https://cdn.usefathom.com/script.js"
+          data-site="LOLUTYWQ"
         />
       )}
     </>
